- add pooling strategies - fix readme
